+++ Sources/MapboxSearch/PublicAPI/FavoriteRecord.swift
@@ -40,6 +40,9 @@ public struct FavoriteRecord: IndexableRecord, SearchResult, Codable, Equatable 
     /// Index in response from server.
     public let serverIndex: Int?
     
+    /// A point accuracy metric for the returned address.
+    public let accuracy: SearchResultAccuracy?
+    
     /// Maki icon name. Use ``icon`` when possible.
     public var iconName: String?
     
@@ -78,6 +81,7 @@ public struct FavoriteRecord: IndexableRecord, SearchResult, Codable, Equatable 
         address: Address?,
         makiIcon: Maki?,
         serverIndex: Int?,
+        accuracy: SearchResultAccuracy?,
         categories: [String]?,
         routablePoints: [RoutablePoint]? = nil,
         resultType: SearchResultType,
@@ -90,6 +94,7 @@ public struct FavoriteRecord: IndexableRecord, SearchResult, Codable, Equatable 
         self.address = address
         self.icon = makiIcon
         self.serverIndex = serverIndex
+        self.accuracy = accuracy
         self.categories = categories
         self.type = resultType
         self.metadata = metadata
@@ -112,6 +117,7 @@ public struct FavoriteRecord: IndexableRecord, SearchResult, Codable, Equatable 
                   address: searchResult.address,
                   makiIcon: searchResult.iconName.flatMap(Maki.init),
                   serverIndex: searchResult.serverIndex,
+                  accuracy: searchResult.accuracy,
                   categories: searchResult.categories,
                   routablePoints: searchResult.routablePoints,
                   resultType: searchResult.type,