+++ Sources/MapboxSearch/PublicAPI/Search Results/SearchResult.swift
@@ -16,6 +16,9 @@ public protocol SearchResult {
     /// Index in response from server.
     var serverIndex: Int? { get }
     
+    /// A point accuracy metric for the returned address.
+    var accuracy: SearchResultAccuracy? { get }
+    
     /// Type of SearchResult. Should be one of address or POI.
     var type: SearchResultType { get }